requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.2.0
openpyxl==3.1.2
xlsxwriter==3.1.2
line-bot-sdk==3.5.1
pillow==10.2.0
psutil==5.9.8
pytz==2025.2
reportlab==4.3.1
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
